pc-agent/requirements.txt
aiomqtt>=2.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
